Decimal math

+++ Calc.py
@@ -1,7 +1,7 @@
 def question():
     print("Continue?")
     i = input("yes = 1 or no = 0?: ")
-    while (int(i) > 0):
+    while (float(i) > 0):
         calculator()
     else:
         print("Bye, bye")
@@ -9,22 +9,22 @@ def question():
 
 def calculator():
     print("Welcome to calculator")
-    a = input("Enter first number: ")
-    b = input("Enter second number: ")
+    a = float(input("Enter first number: "))
+    b = float(input("Enter second number: "))
     c = input("+, -, *, /")
     if c == "+":
-        print(int(a)+int(b))
+        print(a+b)
     if c == "-":
-        print(int(a)-int(b))
+        print(a-b)
     if c == "*":
-        print(int(a)*int(b))
+        print(a*b)
     if c == "/":
-        if int(a) == 0:
+        if a == 0:
             a1010()
-        elif int(b) == 0:
+        elif b == 0:
             a1010()
         else:
-            print(int(a)/int(b))
+            print(a/b)
     question()
